+++ frontend/src/views/contents/Gather/GatherCreate.vue
@@ -3,6 +3,7 @@
         <CreateTitle/>
         <CreateInfo/>
         <CreateDescription/>
+        <CreateBtn/>
     </section>
 </template>
 
@@ -10,6 +11,7 @@
     import CreateTitle from '@/components/body/gather/create/CreateTitle.vue';
     import CreateInfo from '@/components/body/gather/create/CreateInfo.vue';
     import CreateDescription from '@/components/body/gather/create/CreateDescription.vue';
+    import CreateBtn from '@/components/body/gather/create/CreateBtn.vue';
 
 </script>
